pyproject.toml
[tool.poetry]
name = "doe-suite"
version = "0.1.0"
description = ""
authors = ["Nicolas Küchler <nico.kuechler@protonmail.com>"]
# TODO [nku] after poetry update to 1.2.0, we can load .env file via plugin: https://github.com/python-poetry/poetry/issues/337

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
ansible = ">=2.11"
pandas = "^1.3.3"
PyYAML = "^5.4.1"
argparse = "^1.4.0"
PyInputPlus = "^0.2.12"
boto3 = "^1.18.59"
matplotlib = "^3.4.3"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
